Move pages to swollball-specific path
