+++ Pages/Index.cshtml
@@ -40,7 +40,7 @@
             }
         </div>
         <h6>
-            <a href="\lobbylist">Lobby List</a><br>
+            <a href="Swollball\Swollball_LobbyList">Lobby List</a><br>
             <a href="\checklogin">User Info</a><br>
         @Model.BuildNumber
         </h6>
