+++ Pages/Swollball/Swollball_LobbyList.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model Dumdum.LobbyListModel
+@model Dumdum.Pages.Swollball.Swollball_LobbyListModel
 @{
     <table border="1">
         <th>Room ID</th>
@@ -9,7 +9,7 @@
         @foreach (var room in Model.Rooms)
         {
             <tr>
-                <th><a href='/SwollballInsight?RoomId=@room.Item1'>@room.Item1</a></th>
+                <th><a href='/Swollball/SwollballInsight?RoomId=@room.Item1'>@room.Item1</a></th>
                 <th>@room.Item2.ToString()</th>
                 <th>@room.Item3.ToString()</th>
                 <th>@room.Item4.ToString()</th>
